+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project (not yet) adheres to [Semantic Versioning](https://semver.org/spec/
 - Pagination positioning
 - Youtube embedding (dsvgo repo)
 - Padding of copyright text in ProjectTile
+- Kiezradar map district borders and masking
 
 ## mB-v2602.3